Move FeatureEditorFormView.
